Fix CI: ProductCard test Product type and clippy warnings in contracts.
Add required description field to test mock and clean up verifier imports and integrity anchor cast.

+++ smart-contract/contracts/src/storage.rs
@@ -443,7 +443,7 @@ pub fn add_product_anchor(env: &Env, product_id: &String, anchor_id: u64) -> Res
         .get(&key)
         .unwrap_or_else(|| Vec::new(env));
 
-    if ids.len() >= crate::integrity_anchor::MAX_ANCHORS_PER_PRODUCT as u32 {
+    if ids.len() >= crate::integrity_anchor::MAX_ANCHORS_PER_PRODUCT {
         return Err(Error::TooManyAnchors);
     }
